requirements.txt
requests
PyPDF2
black
pandas
scikit-learn
matplotlib
faiss-cpu
faiss-gpu
pytorch